split loader
